Create 02_first_bit
Write a program that prints the bit at position 1 of a number.
Examples
 
Input	Output	   
2	1	   
51	1	   
13	0	   
24	0	 
Hints
Declare two variables (n and bitAtPosition1). 
Read the user input from the console. 
Find the value of the bit at position 1 (position 1 is the second bit from right to left: [7, 6, 5, 4, 3, 2, 1, 0]): 
Shift the number n times to the right (where n is the position, in this case it is 1) by using the >> operator. In that way the bit we want to check will be at position 0;
Find the bit at position 0. Use & 1 operator expression to extract the value of a bit. By using the following formulae () you check whether the bit at position 0 is equal to 1 or not. If the bit is equal to 1 the result is 1 if the bit is not equal - the result is 0;
Save the result in bitAtPosition1;
Print the result on the console.