Create _03_p_th_bit
Write a program that prints the bit at position p of a number.
Examples
 
Input	Output	   
2145
5	1	   
512
0	0	   
111
8	0	   
255
7	1	 
Hints
Declare three variables (n, p and bitAtPositionP). 
Read the user input from the console. 
Find the value of the bit at position p: 
Shift the number p times to the right (where p is the position) by using the >> operator. In that way the bit we want to check will be at position 0;
Find the bit at position 0. Use & 1 operator expression to extract the value of a bit. By using the following formula (bitAtPositionP & 1) you check whether the bit at position 0 is equal to 1 or not. If the bit is equal to 1 the result is 1 if the bit is not equal - the result is 0;
Save the result in bitAtPosition1;