Create dem03_Last_Stop
Create a program that follows instructions in order to fulfil a quest. First, you will receive a collection of numbers – each representing a painting number. After that, you are going to be receiving instructions, until the "END" command is given.
Change {paintingNumber} {changedNumber} – find the painting with the first number in the collection (if it exists) and change its number with the second number – {changedNumber}.
Hide {paintingNumber} – find the painting with this value and if it exists and hide it (remove it).
Switch {paintingNumber} {paintingNumber2} – find the given paintings in the collections if they exist and switch their places.
Insert {place} {paintingNumber} – insert the painting (paintingNumber) on the next place after the given one, if it exists.
Reverse – you must reverse the order of the paintings.
Once you complete the instructions, print the numbers of the paintings on a single line, split by a space.
Input / Constraints
On the 1st line, you are going to receive the numbers of the paintings, split by a single space – integer numbers in the range [1…1000]
On the next lines, you are going to receive commands, until you receive the "END" command
Output
Print the message you have received after the conversion of all numbers on a single line